Fix clippy lints + missed aetna-web HostDiagnostics call site
The clippy CI job (workspace + wasm, -D warnings) had been failing
alongside rustfmt — also drift from the color port:
  - excessive-precision float literals in color/convert.rs matrices
    (truncated to canonical f32; bit-identical values)
  - f32→f32 self-casts in theme/style.rs + the vulkano/ash demo crates
  - derivable Default impl on ColorManagementStatus
  - doc_lazy_continuation + collapsible_match in wayland_color.rs

Also fixes a real miss: aetna-web's WebGPU host builds its own
HostDiagnostics and wasn't updated when working_color_space +
color_management were added, so the wasm target failed to compile.
The web host gets ColorManagementStatus::Unavailable + the default
sRGB-linear working space (it doesn't negotiate color management).

+++ crates/aetna-core/src/theme/style.rs
@@ -403,7 +403,7 @@ fn text_on_solid(c: Color) -> Color {
         _ => {}
     }
 
-    let lum = 0.299 * c.r as f32 + 0.587 * c.g as f32 + 0.114 * c.b as f32;
+    let lum = 0.299 * c.r + 0.587 * c.g + 0.114 * c.b;
     if lum > 150.0 {
         Color::srgb_u8a(8, 16, 25, 255)
     } else {
